+++ crates/move-bytecode-to-wasm/src/translation/bytecodes/enums.rs
@@ -83,7 +83,7 @@ pub fn pack_variant(
 
                         // Create a pointer for the value
                         builder
-                            .i32_const(data_size as i32)
+                            .i32_const(data_size)
                             .call(compilation_ctx.allocator)
                             .local_tee(ptr_to_data);
